BirthdayBox/requirements.txt
google-auth-httplib2
google-api-python-client
python-dotenv
